# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "CAB BOOKING",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/1.0.0",
      "description": "version-prefix"
    },
    {
      "url": "http://localhost:8000/1.0.0",
      "description": "development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Cab": {
        "type": "object",
        "properties": {
          "cab_info": {
            "type": "object",
            "properties": {
              "cab_number": {
                "type": "string",
                "example": "MH11-CE-8226"
              },
              "driver_name": {
                "type": "string",
                "example": "deepak"
              },
              "driver_mobile": {
                "type": "string",
                "example": "8888225544"
              }
            }
          },
          "current_latitude": {
            "type": "number",
            "example": 133333333
          },
          "current_longitude": {
            "type": "number",
            "example": 122222222
          }
        }
      },
...
    },
    "parameters": {
      "limit": {
        "in": "query",
        "name": "limit",
        "description": "Number of objects returned",
        "schema": {
          "type": "integer"
        },
        "example": 5
      },
      "offset": {
        "name": "offset",
        "in": "query",
        "description": "Offset of query",
        "schema": {
          "type": "integer"
        },
        "example": 1
      },
      "orderBy": {
        "name": "order_by",
        "in": "query",
        "description": "Field to sort by",
        "schema": {
          "type": "string"
        },
        "example": "id"
      },
      "order": {
        "name": "order",
        "in": "query",
        "description": "Sort records in ASC or DESC",
        "schema": {
          "type": "string"
        },
        "example": "DESC"
      },
      "id": {
        "name": "id",
        "in": "query",
        "description": "Array of ids to retrieve",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "example": [
          1,
          2,
          3
        ]
      }
    },
...
  },
  "paths": {
...
    "/cab/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "description": "The id of the Cab to perform action on.",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Cab"
        ],
        "summary": "Get Cab by id",
        "description": "Get Cab by id",
        "operationId": "getCabById",
        "security": [
          {
            "BearerAuth": [
              "default"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Cab"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Cab not found"
          }
        }
      },
      "put": {
        "tags": [
          "Cab"
        ],
        "summary": "Update Cab",
        "description": "Update Cab",
        "operationId": "updateCab",
        "security": [
          {
            "BearerAuth": [
              "default"
            ]
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/components/schemas/Cab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Cab"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Cab not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Cab"
        ],
        "summary": "Delete Cab",
        "description": "Delete Cab",
        "operationId": "deleteCab",
        "security": [
          {
            "BearerAuth": [
              "default"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Cab not found"
          }
        }
      }
    },
...
    "/find_my_nearest_cab": {
      "get": {
        "tags": [
          "My Nearest Cab"
        ],
        "summary": "Get My Nearest Cab list",
        "description": "Get My Nearest Cab list.",
        "operationId": "getFindMyNearestCabList",
        "security": [
          {
            "BearerAuth": [
              "default"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/orderBy"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "in": "query",
            "name": "latitude",
            "description": "Find cab by user current latitude",
            "schema": {
              "type": "number"
            },
            "example": 133333333
          },
          {
            "in": "query",
            "name": "longitude",
            "description": "Find cab by user current longitude",
            "schema": {
              "type": "number"
            },
            "example": 122222222
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "allOf": [
                      {
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          }
                        }
                      },
                      {
                        "$ref": "#/components/schemas/Cab"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}